refactor: EditableCommentItem 컴포넌트와 CommentItem 컴포넌트 하나의 컴포넌트로 합침
