+++ src/app/post-detail/_components/CommentList.tsx
@@ -1,5 +1,4 @@
 import { CommentType } from '@/types/common';
-import EditableCommentItem from './EditableCommentItem';
 import { SupabaseServerApi } from '@/api/supabase.server.api';
 import CommentItem from './CommentItem';
 
@@ -16,15 +15,11 @@ async function CommentList({ commentList }: Props) {
         const isOwnedByLoginUser =
           user && user.id === commentItem.user_id ? true : false;
 
-        return isOwnedByLoginUser ? (
-          <EditableCommentItem
-            key={commentItem.id}
-            commentItem={commentItem}
-          />
-        ) : (
+        return (
           <CommentItem
             key={commentItem.id}
             commentItem={commentItem}
+            isOwnedByLoginUser={isOwnedByLoginUser}
           />
         );
       })}
